requirements.txt
openai
sqlalchemy
sqlparse
pymysql
python-dotenv
groq